Crate consensus_core

Macros

bail

ensure

Structs

BlockRef

BlockRef uniquely identifies a VerifiedBlock via digest. It also contains the slot info (round and author) so it can be used in logic such as aggregating stakes for a round.

Clock

A clock that allows to derive the current UNIX system timestamp while guaranteeing that timestamp will be monotonically incremented, tolerating ntp and system clock changes and corrections. Explicitly avoid to make [Clock] cloneable to ensure that a single instance is shared behind an [Arc] wherever is needed in order to make sure that consecutive calls to receive the system timestamp will remain monotonically increasing.

CommitConsumer

CommitConsumerMonitor

CommitDigest

Digest of a consensus commit.

CommitRef

Uniquely identifies a commit with its index and digest.

CommittedSubDag

The output of consensus to execution is an ordered list of CommittedSubDag. Each CommittedSubDag contains the information needed to execution transactions in the consensus commit.

TestBlock

Exported API for testing. This struct is public for testing in other crates.

Transaction

Exported API for testing. IOTA transaction in serialised bytes

TransactionClient

VerifiedBlock

Exported API for testing. VerifiedBlock allows full access to its content. Note: clone() is relatively cheap with most underlying data refcounted.

Enums

BlockStatus

ClientError

ConsensusAuthority

Exported consensus API. ConsensusAuthority is used by IOTA to manage the lifetime of AuthorityNode. It hides the details of the implementation from the caller, MysticetiManager.

ValidationError

Traits

BlockAPI

TransactionVerifier

TransactionVerifier implementation is supplied by IOTA to validate transactions in a block, before acceptance of the block.

Functions

to_socket_addr

Attempts to convert a multiaddr of the form /[ip4,ip6]/{}/[udp,tcp]/{port} into a SocketAddr value.

Type Aliases

BlockTimestampMs

Exported API for testing. Block proposal as epoch UNIX timestamp in milliseconds.

CommitIndex

Index of a commit among all consensus commits.

Round

Round number of a block.