Trait consensus_core::TransactionVerifier

pub trait TransactionVerifier: Send + Sync + 'static {
    // Required methods
    fn verify_batch(&self, batch: &[&[u8]]) -> Result<(), ValidationError>;
    fn verify_and_vote_batch<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        batch: &'life1 [&'life2 [u8]],
    ) -> Pin<Box<dyn Future<Output = Result<Vec<TransactionIndex>, ValidationError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
}

TransactionVerifier implementation is supplied by Sui to validate transactions in a block, before acceptance of the block.

Required Methods

fn verify_batch(&self, batch: &[&[u8]]) -> Result<(), ValidationError>

Determines if this batch of transactions is valid. Fails if any one of the transactions is invalid.

fn verify_and_vote_batch<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, batch: &'life1 [&'life2 [u8]], ) -> Pin<Box<dyn Future<Output = Result<Vec<TransactionIndex>, ValidationError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Returns indices of transactions to reject, validator error over transactions. Currently only uncertified user transactions can be rejected. The rest of transactions are implicitly voted to be accepted. When the result is an error, the whole block should be rejected from local DAG instead.