Crate consensus_core

Macros

bail

ensure

Structs

CommitConsumer

CommitDigest

Digest of a consensus commit.

CommitRef

Uniquely identifies a commit with its index and digest.

CommittedSubDag

The output of consensus to execution is an ordered list of CommittedSubDag. Each CommittedSubDag contains the information needed to execution transactions in the consensus commit.

TransactionClient

Enums

ConsensusAuthority

ConsensusAuthority is used by Sui to manage the lifetime of AuthorityNode. It hides the details of the implementation from the caller, MysticetiManager.

ValidationError

Traits

BlockAPI

TransactionVerifier

TransactionVerifier implementation is supplied by Sui to validate transactions in a block, before acceptance of the block.

Type Aliases

CommitIndex

Index of a commit among all consensus commits.

Round

Round number of a block.